8. Стороны прямоугольника. [#137735 #415801]

+++ src/test/java/ru/job4j/condition/SqAreaTest.java
@@ -11,7 +11,7 @@ public class SqAreaTest {
     public void whenP6K2Square2() {
         int expected = 2;
         int p = 6;
-        double k = 2;
+        int k = 2;
         double out = SqArea.square(p, k);
         Assert.assertEquals(expected, out, 0.01);
     }
